+++ tools/c7n_left/c7n_left/entry.py
@@ -1,3 +1,3 @@
 def initialize_iac():
     # import to get side effect registration into clouds
-    from .provider import TerraformProvider  # noqa
+    from .providers.terraform import TerraformProvider  # noqa
