--- tools/c7n_left/c7n_left/provider.py
@@ -1,290 +1,0 @@
-# Copyright The Cloud Custodian Authors.
-# SPDX-License-Identifier: Apache-2.0
-#
-import fnmatch
-import logging
-
-from c7n.actions import ActionRegistry
-from c7n.cache import NullCache
-from c7n.filters import FilterRegistry
-from c7n.manager import ResourceManager
-from c7n.policy import execution
-
-from c7n.provider import Provider, clouds
-from c7n.policy import PolicyExecutionMode
-from c7n.utils import type_schema
-
-from tfparse import load_from_path
-
-
-log = logging.getLogger("c7n.iac")
-
-
-class IACSourceProvider(Provider):
-
-    display_name = "IAC"
-
-    def get_session_factory(self, options):
-        return lambda *args, **kw: None
-
-    def initialize(self, options):
-        pass
-
-    def initialize_policies(self, policies, options):
-        return policies
-
-
-class CollectionRunner:
-    def __init__(self, policies, options, reporter):
-        self.policies = policies
-        self.options = options
-        self.reporter = reporter
-
-    def run(self):
-        event = self.get_event()
-        provider = self.get_provider()
-
-        if not provider.match_dir(self.options.source_dir):
-            raise NotImplementedError(
-                "no %s source files found" % provider.provider_name
-            )
-
-        graph = provider.parse(self.options.source_dir)
-
-        for p in self.policies:
-            p.expand_variables(p.get_variables())
-            p.validate()
-
-        self.reporter.on_execution_started(self.policies)
-        # consider inverting this order to allow for results grouped by policy
-        # at the moment, we're doing results grouped by resource.
-        for rtype, resources in graph.get_resources_by_type():
-            for p in self.policies:
-                if not self.match_type(rtype, p):
-                    continue
-                result_set = self.run_policy(p, graph, resources, event)
-                if result_set:
-                    self.reporter.on_results(result_set)
-        self.reporter.on_execution_ended()
-
-    def run_policy(self, policy, graph, resources, event):
-        event = dict(event)
-        event.update({"graph": graph, "resources": resources})
-        return policy.push(event)
-
-    def get_provider(self):
-        provider_name = {p.provider_name for p in self.policies}.pop()
-        provider = clouds[provider_name]()
-        return provider
-
-    def get_event(self):
-        return {"config": self.options}
-
-    @staticmethod
-    def match_type(rtype, p):
-        if isinstance(p.resource_type, str):
-            return fnmatch.fnmatch(rtype, p.resource_type.split(".", 1)[-1])
-        if isinstance(p.resource_type, list):
-            for pr in p.resource_type:
-                return fnmatch.fnmatch(rtype, pr.split(".", 1)[-1])
-
-
-class IACSourceMode(PolicyExecutionMode):
-    @property
-    def manager(self):
-        return self.policy.resource_manager
-
-    def run(self, event, ctx):
-        if not self.policy.is_runnable(event):
-            return []
-
-        resources = event["resources"]
-        resources = self.manager.filter_resources(resources, event)
-        return self.as_results(resources)
-
-    def as_results(self, resources):
-        return ResultSet([PolicyResourceResult(r, self.policy) for r in resources])
-
-
-class ResultSet(list):
-    pass
-
-
-class PolicyResourceResult:
-    def __init__(self, resource, policy):
-        self.resource = resource
-        self.policy = policy
-
-
-class IACResourceManager(ResourceManager):
-
-    filter_registry = FilterRegistry("iac.filters")
-    action_registry = ActionRegistry("iac.actions")
-    log = log
-
-    def __init__(self, ctx, data):
-        self.ctx = ctx
-        self.data = data
-        self._cache = NullCache(None)
-        self.session_factory = lambda: None
-        self.filters = self.filter_registry.parse(self.data.get("filters", []), self)
-        self.actions = self.action_registry.parse(self.data.get("actions", []), self)
-
-    def get_resource_manager(self, resource_type, data=None):
-        return self.__class__(self.ctx, data or {})
-
-
-class IACResourceMap(object):
-
-    resource_class = None
-
-    def __init__(self, prefix):
-        self.prefix = prefix
-
-    def __contains__(self, k):
-        if k.startswith(self.prefix):
-            return True
-        return False
-
-    def __getitem__(self, k):
-        if k.startswith(self.prefix):
-            return self.resource_class
-        raise KeyError(k)
-
-    def __iter__(self):
-        return iter(())
-
-    def notify(self, *args):
-        pass
-
-    def keys(self):
-        return ()
-
-    def items(self):
-        return ()
-
-    def get(self, k, default=None):
-        # that the resource is in the map has alerady been verified
-        # we get the unprefixed resource on get
-        return self.resource_class
-
-
-class TerraformResourceManager(IACResourceManager):
-    pass
-
-
-class TerraformResourceMap(IACResourceMap):
-
-    resource_class = TerraformResourceManager
-
-
-@clouds.register("terraform")
-class TerraformProvider(IACSourceProvider):
-
-    display_name = "Terraform"
-    resource_prefix = "terraform"
-    resource_map = TerraformResourceMap(resource_prefix)
-    resources = resource_map
-
-    def initialize_policies(self, policies, options):
-        for p in policies:
-            p.data["mode"] = {"type": "terraform-source"}
-        return policies
-
-    def parse(self, source_dir):
-        graph = TerraformGraph(load_from_path(source_dir), source_dir)
-        log.debug("Loaded %d resources", len(graph))
-        return graph
-
-    def match_dir(self, source_dir):
-        files = list(source_dir.glob("*.tf"))
-        files += list(source_dir.glob("*.tf.json"))
-        return files
-
-
-@execution.register("terraform-source")
-class TerraformSource(IACSourceMode):
-
-    schema = type_schema("terraform-source")
-
-
-class TerraformResource(dict):
-
-    __slots__ = ("name", "data", "location")
-
-    # pygments lexer
-    format = "terraform"
-
-    def __init__(self, name, data):
-        self.name = name
-        if isinstance(data["__tfmeta"], list):
-            self.location = data["__tfmeta"][0]
-        else:
-            self.location = data["__tfmeta"]
-        super().__init__(data)
-
-    @property
-    def filename(self):
-        return self.location["filename"]
-
-    @property
-    def line_start(self):
-        return self.location["line_start"]
-
-    @property
-    def line_end(self):
-        return self.location["line_end"]
-
-    @property
-    def src_dir(self):
-        return self.location["src_dir"]
-
-    def get_source_lines(self):
-        lines = (self.src_dir / self.filename).read_text().split("\n")
-        return lines[self.line_start - 1 : self.line_end]  # noqa
-
-
-class ResourceGraph:
-    def __init__(self, resource_data, src_dir):
-        self.resource_data = resource_data
-        self.src_dir = src_dir
-
-    def get_resource_by_type(self):
-        raise NotImplementedError()
-
-
-class TerraformGraph(ResourceGraph):
-    def __len__(self):
-        return sum(map(len, self.resource_data.values()))
-
-    def get_resources_by_type(self, types=()):
-        if isinstance(types, str):
-            types = (types,)
-        for type_name, type_items in self.resource_data.items():
-            if types and type_name not in types:
-                continue
-            if type_name == "data":
-                for data_type, data_items in type_items.items():
-                    resources = []
-                    for name, data in data_items.items():
-                        resources.append(self.as_resource(name, data))
-                    yield "%s.%s" % (type_name, data_type), resources
-            elif type_name == "moved":
-                yield type_name, self.as_resource(type_name, data)
-            elif type_name == "locals":
-                yield type_name, self.as_resource(type_name, data)
-            elif type_name == "terraform":
-                yield type_name, self.as_resource(type_name, data)
-            else:
-                resources = []
-                for name, data in type_items.items():
-                    resources.append(self.as_resource(name, data))
-                yield type_name, resources
-
-    def as_resource(self, name, data):
-        if isinstance(data["__tfmeta"], list):
-            for m in data["__tfmeta"]:
-                m["src_dir"] = self.src_dir
-        else:
-            data["__tfmeta"]["src_dir"] = self.src_dir
-        return TerraformResource(name, data)
